gym reservation bot/requirements.txt
selenium>=4.15.0
google-cloud-secret-manager>=2.16.0

